src/requirements.txt
stomp.py>=8.0.0
PyXB~=1.2.6
kafka-python==1.4.6
confluent_kafka
requests
pyspark==3.3.1
avro
faust
fastavro
cassandra-driver
streamlit